+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.3 (Upcoming)
  - Improved MicroPython flasher UI
  - Seperated logic for espressif devices in flasher to a seperate file
+ - Added beta download as ZIP functionality
 
 ## 1.1.2
  - Added ability to use arrow keys to select command shown in command palette (`CTRL+SHIFT+P`)